requirements.txt
meshtastic>=2.0.0
pypubsub>=4.0.3
ollama>=0.1.0
python-dotenv>=0.19.0
